Pre/PostCondition and NQueen fixes

+++ Project7/Project7/src/NQueen.cpp
@@ -184,7 +184,7 @@ void NQueen::displayBoard() const
     printf("%s", ss.str().c_str());
 }
 
-int NQueen::getSize() const
+size_t NQueen::getSize() const
 {
     return m_BoardSize;
 }
